refactor(frontend): move Solana fee arithmetic out of the compute budget instruction parser
The sol-instructions-*.utils.ts files are per-program instruction parsers, each
exporting a single parseSol...Instruction. Fee maths does not belong there, so it
moves to $sol/utils/fee.utils.ts, mirroring $eth/utils/fee.utils.ts.
